pyproject.toml
[project]
name = "person-up-ops-project"
version = "0.1.0"
description = "Person Up Ops Project backend scaffold"
requires-python = ">=3.13,<3.14"
dependencies = [
  "fastapi>=0.115,<1.0",
  "uvicorn>=0.34,<1.0",
  "pydantic-settings>=2.8,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "alembic>=1.16,<2.0",
  "celery>=5.5,<6.0",
  "redis>=5.2,<6.0",
  "psycopg[binary]>=3.2,<4",
]

[dependency-groups]
dev = [
  "ruff>=0.11,<1.0",
]
test = [
  "httpx>=0.28,<1.0",
]

[tool.uv]
package = false
